SFC: add commodity exchange shortcut buttons to the destination field
Adds ANT / BEN / HRT / MOR buttons inline at the right of the SFC
destination address input, so a ship can be sent to a commodity exchange
without typing the station name. The input flexes to make room; nothing
else in the buffer layout moves.

Each button types the station name into the AddressSelector and picks the
matching autosuggest entry — the same flow the user would perform by
hand, kept behind an explicit click so no server request is automated.

Extracts the address-selection helper (previously private to the CONTD
paste-import feature) to
src/infrastructure/prun-ui/utils/select-address.ts, and the waitFor poll
helper to src/utils/wait-for.ts, so the new feature reuses them instead
of adding a third copy.
