+++ src/features/basic/sfc-exchange-destinations.module.css
@@ -1,0 +1,25 @@
+/* The address selector is the input's own react-autosuggest wrapper — the
+   suggestion list lives in #autosuggest-portal, so laying it out as a row only
+   affects the input and the appended buttons. */
+.container {
+  display: flex;
+  align-items: center;
+  gap: 3px;
+}
+
+.input {
+  flex: 1 1 auto;
+  min-width: 0;
+}
+
+.buttons {
+  display: flex;
+  flex: 0 0 auto;
+  gap: 3px;
+}
+
+.button {
+  margin: 0;
+  padding-left: 5px;
+  padding-right: 5px;
+}
